+++ src/JanelaPrincipal.java
@@ -1,7 +1,47 @@
+import java.awt.BorderLayout;
+import java.awt.Font;
 
-public class JanelaPrincipal {
+import javax.swing.JButton;
+import javax.swing.JFrame;
+import javax.swing.JLabel;
+import javax.swing.JPanel;
+import javax.swing.SwingUtilities;
+
+public class JanelaPrincipal extends JFrame{
+
+    
+        SwingUtilities.invokeLater(() -> {
+        JanelaPrincipal janela = new JanelaPrincipal();
+        janela.setVisible(true);
 
     public void setVisible(boolean b) {
     }
 
+    public JanelaPrincipal{
+
+        setTitle("Jogo Super Trunfo");
+        setSize(800, 600);
+        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
+        setLocationRelativeTo(null); 
+
+
+        JLabel tituloLabel = new JLabel("Bem-vindo ao Jogo Super Trunfo!");
+        tituloLabel.setFont(new Font("Arial", Font.BOLD, 24));
+        tituloLabel.setHorizontalAlignment(JLabel.CENTER);
+        tituloLabel.setVerticalAlignment(JLabel.CENTER);
+
+        JButton iniciarButton = new JButton("Iniciar Jogo");
+        iniciarButton.setFont(new Font("Arial", Font.PLAIN, 18));
+        iniciarButton.addActionListener(e -> iniciarJogo());
+
+
+        JPanel painelPrincipal = new JPanel(new BorderLayout());
+        painelPrincipal.add(tituloLabel, BorderLayout.CENTER);
+        painelPrincipal.add(iniciarButton, BorderLayout.SOUTH);
+
+        add(painelPrincipal);
+
+
+    });
+    }
 }
