+++ src/Principal.java
@@ -1,49 +1,19 @@
-import java.awt.BorderLayout;
-import java.awt.Font;
 
-import javax.swing.JButton;
-import javax.swing.JFrame;
-import javax.swing.JLabel;
-import javax.swing.JPanel;
-import javax.swing.SwingUtilities;
 
-public class Principal extends JFrame{
+public class Principal{
     public Principal() {
 
-        setTitle("Jogo Super Trunfo");
-        setSize(800, 600);
-        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setLocationRelativeTo(null); 
 
-
-        JLabel tituloLabel = new JLabel("Bem-vindo ao Jogo Super Trunfo!");
-        tituloLabel.setFont(new Font("Arial", Font.BOLD, 24));
-        tituloLabel.setHorizontalAlignment(JLabel.CENTER);
-        tituloLabel.setVerticalAlignment(JLabel.CENTER);
-
-        JButton iniciarButton = new JButton("Iniciar Jogo");
-        iniciarButton.setFont(new Font("Arial", Font.PLAIN, 18));
-        iniciarButton.addActionListener(e -> iniciarJogo());
-
-
-        JPanel painelPrincipal = new JPanel(new BorderLayout());
-        painelPrincipal.add(tituloLabel, BorderLayout.CENTER);
-        painelPrincipal.add(iniciarButton, BorderLayout.SOUTH);
-
-        add(painelPrincipal);
     }
     private Object iniciarJogo() {
         return null;
     }
     public static void main(String[] args) {
         
-            	SwingUtilities.invokeLater(() -> {
-        	JanelaPrincipal janela = new JanelaPrincipal();
-           	janela.setVisible(true);
-      		});
-        
+
         Jogo jogo = new Jogo();
         jogo.jogar();
+
     }
     @Override
     public String toString() {
